Implemented constraint solver Implemented constraint solver, test with pendulum. There appears to be numerical errors. Errors decrease with step size so implementing RK4 solver may fix this problem (Or perhaps there is something else going on?). Preparing for rewrite of ndsolver for implementation of RK4 method.

+++ src/constraint.h
@@ -55,7 +55,7 @@ namespace physics{
     class ConstraintPivot1P : public Constraint{
         public:
 
-            ConstraintPivot1P(Particle p, long double _l, int id) : Constraint({id}){
+            ConstraintPivot1P(Particle p, double _l, int id) : Constraint({id}){
                 l = _l;
             }
